@@ -1,9 +1,6 @@
 ﻿using System;
 using System.IO;
 using System.IO.Compression;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Server.RevealSDK;
 using Reveal.Sdk;
 using Microsoft.AspNetCore.Builder;
@@ -11,13 +8,10 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.ResponseCompression;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Swashbuckle.AspNetCore.SwaggerGen;
-using Microsoft.OpenApi.Models;
-using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc.Controllers;
+using Microsoft.Extensions.Hosting;
 
 namespace Server
 {
@@ -25,7 +19,7 @@ namespace Server
     {
         private string _webRootPath;
 
-        public Startup(IConfiguration configuration, IHostingEnvironment env)
+        public Startup(IConfiguration configuration, IWebHostEnvironment env)
         {
             Configuration = configuration;
             _webRootPath = env.WebRootPath;
@@ -49,26 +43,13 @@ namespace Server
                 }; 
             });
 
-            services.Configure<CookiePolicyOptions>(options =>
-            {
-                // This lambda determines whether user consent for non-essential cookies is needed for a given request.
-                options.CheckConsentNeeded = context => true;
-                options.MinimumSameSitePolicy = SameSiteMode.None;
-            });
-
-            services.Configure<GzipCompressionProviderOptions>(options => options.Level = CompressionLevel.Fastest);
-            services.AddResponseCompression(options =>
-            {
-                options.Providers.Add<GzipCompressionProvider>();
-                options.EnableForHttps = true;
-            });
             services.AddRevealServices(new RevealEmbedSettings
             {
                 CachePath = @"C:\Temp2",
                 LocalFileStoragePath = Path.Combine(_webRootPath, "App_Data", "RVLocalFiles")
-            }, new RevealSdkContext(_webRootPath), new RevealUserContext());
+            }, new RevealSdkContext(_webRootPath));
 
-            services.AddMvc().AddReveal().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddControllers().AddReveal();
         }
 
         protected virtual RevealSdkContextBase CreateSdkContext()
@@ -77,17 +58,11 @@ namespace Server
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Home/Error");
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
             }
 
             app.UseSwagger();
@@ -98,11 +73,10 @@ namespace Server
             });
 
             app.UseHttpsRedirection();
-            app.UseCors(x => x.AllowAnyOrigin().AllowAnyMethod().AllowAnyHeader());
-            app.UseCookiePolicy();
+            app.UseRouting();
 
-            app.UseResponseCompression();
-            //app.UseStaticFiles();
+            app.UseCors(x => x.AllowAnyOrigin().AllowAnyMethod().AllowAnyHeader());
+
             app.UseStaticFiles(new StaticFileOptions
             {
                 OnPrepareResponse = context =>
@@ -116,11 +90,9 @@ namespace Server
                 }
             });
 
-            app.UseMvc(routes =>
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
-                    name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                endpoints.MapControllers();
             });
         }
     }
